+++ src/main/java/org/folio/search/utils/SearchConverterUtils.java
@@ -3,9 +3,7 @@ package org.folio.search.utils;
 import static java.util.Collections.emptyMap;
 import static org.apache.commons.collections4.MapUtils.getString;
 import static org.apache.commons.lang3.BooleanUtils.isTrue;
-import static org.folio.search.domain.dto.ResourceEventType.UPDATE;
 import static org.folio.search.utils.SearchUtils.ID_FIELD;
-import static org.folio.search.utils.SearchUtils.SOURCE_CONSORTIUM_PREFIX;
 import static org.folio.search.utils.SearchUtils.SOURCE_FIELD;
 
 import java.util.Arrays;
@@ -17,7 +15,6 @@ import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.collections4.MapUtils;
-import org.apache.commons.lang3.Strings;
 import org.folio.search.domain.dto.ResourceEvent;
 import org.springframework.stereotype.Component;
 
@@ -184,14 +181,6 @@ public class SearchConverterUtils {
         target.put(field, source.get(field));
       }
     }
-  }
-
-  public static boolean isUpdateEventForResourceSharing(ResourceEvent event) {
-    var newSource = getResourceSource(getNewAsMap(event));
-    return event.getType() == UPDATE
-           && Strings.CS.startsWith(newSource, SOURCE_CONSORTIUM_PREFIX)
-           && Objects.equals(getResourceSource(getOldAsMap(event)),
-      Strings.CS.removeStart(newSource, SOURCE_CONSORTIUM_PREFIX));
   }
 
   /**